updating github from foresmo dev svn repo

+++ source/solar/Solar/Cli/MakeVendor.php
@@ -12,13 +12,13 @@
  * 
  * @license http://opensource.org/licenses/bsd-license.php BSD
  * 
- * @version $Id: MakeVendor.php 4038 2009-09-18 01:18:31Z pmjones $
+ * @version $Id: MakeVendor.php 4490 2010-03-02 15:51:00Z pmjones $
  * 
  * @todo Make Vendor_App_Hello, Vendor_Cli_Help.  Also make Vendor_App_Base
  * and Vendor_Cli_Base?
  * 
  */
-class Solar_Cli_MakeVendor extends Solar_Cli_Base
+class Solar_Cli_MakeVendor extends Solar_Controller_Command
 {
     /**
      * 
@@ -46,19 +46,29 @@ class Solar_Cli_MakeVendor extends Solar_Cli_Base
      * 
      */
     protected $_dirs = array(
-        '/{:dashes}/script',
+        '/{:dashes}/config',
         '/{:dashes}/docs',
+        '/{:dashes}/script',
         '/{:dashes}/tests',
+        '/{:dashes}/tests/Fixture',
+        '/{:dashes}/tests/Fixture/{:studly}',
+        '/{:dashes}/tests/Mock',
+        '/{:dashes}/tests/Mock/{:studly}',
         '/{:dashes}/tests/Test',
         '/{:dashes}/tests/Test/{:studly}',
-        '/{:dashes}/{:studly}/App/Public',
-        '/{:dashes}/{:studly}/Model',
+        '/{:dashes}/{:studly}/Cli/Help',
+        '/{:dashes}/{:studly}/Cli/Help/Info',
+        '/{:dashes}/{:studly}/Cli/Help/Locale',
+        '/{:dashes}/{:studly}/Controller/Bread/Locale',
+        '/{:dashes}/{:studly}/Controller/Bread/Public',
+        '/{:dashes}/{:studly}/Controller/Bread/View',
+        '/{:dashes}/{:studly}/Controller/Command/Info',
+        '/{:dashes}/{:studly}/Controller/Command/Locale',
         '/{:dashes}/{:studly}/Controller/Page/Layout',
         '/{:dashes}/{:studly}/Controller/Page/Locale',
+        '/{:dashes}/{:studly}/Controller/Page/Public',
         '/{:dashes}/{:studly}/Controller/Page/View',
-        '/{:dashes}/{:studly}/Controller/Model/Layout',
-        '/{:dashes}/{:studly}/Controller/Model/Locale',
-        '/{:dashes}/{:studly}/Controller/Model/View',
+        '/{:dashes}/{:studly}/Model',
     );
     
     /**
@@ -83,7 +93,7 @@ class Solar_Cli_MakeVendor extends Solar_Cli_Base
     {
         // we need a vendor name, at least
         if (! $vendor) {
-            throw $this->_exception('ERR_NO_VENDOR_NAME');
+            throw $this->_exception('ERR_NO_VENDOR');
         }
         
         // build "foo-bar" and "FooBar" versions of the vendor name.
@@ -95,21 +105,6 @@ class Solar_Cli_MakeVendor extends Solar_Cli_Base
         $this->_createDirs();
         $this->_createFiles();
         $this->_createLinks();
-        
-        // done!
-        $this->_outln("Done!");
-        
-        $this->_outln(
-                "Remember to add '{$this->_studly}_App' to the "
-              . "['Solar_Controller_Front']['classes'] element "
-              . "in your config file so that it finds your apps."
-        );
-
-        $this->_outln(
-                "Remember to add '{$this->_studly}_Model' to the "
-              . "['Solar_Sql_Model_Catalog']['classes'] element "
-              . "in your config file so that it finds your models."
-        );
     }
     
     /**
@@ -150,58 +145,8 @@ class Solar_Cli_MakeVendor extends Solar_Cli_Base
      */
     protected function _createLinks()
     {
-        $this->_outln('Making links.');
-        
-        $links = array(
-            
-            // include/Vendor -> ../source/vendor/Vendor
-            array(
-                'dir' => "include",
-                'tgt' => $this->_studly,
-                'src' => "../source/{$this->_dashes}/$this->_studly",
-            ),
-            
-            // include/Test/Vendor => ../../source/vendor/tests/Test/Vendor
-            array(
-                'dir' => "include/Test",
-                'tgt' => $this->_studly,
-                'src' => "../../source/{$this->_dashes}/tests/Test/$this->_studly",
-            ),
-            
-            // docroot/public/Vendor -> ../../include/Vendor/App/Public
-            array(
-                'dir' => "docroot/public",
-                'tgt' => $this->_studly,
-                'src' => "../../include/{$this->_studly}/App/Public",
-            ),
-            
-            // script/vendor -> ../source/solar/script/solar
-            array(
-                'dir' => "script",
-                'tgt' => $this->_dashes,
-                'src' => "../source/solar/script/solar",
-            ),
-        );
-        
-        $system = Solar::$system;
-        foreach ($links as $link) {
-            
-            // $dir, $src, $tgt
-            extract($link);
-            
-            // skip it?
-            $link = "$dir/$tgt";
-            if (file_exists("$system/$link")) {
-                $this->_outln("Link $link exists.");
-                continue;
-            }
-            
-            // make it
-            $this->_out("Making link $link ... ");
-            $cmd = "cd $system/$dir; ln -s $src $tgt";
-            passthru($cmd);
-            $this->_outln("done.");
-        }
+        $link_vendor = Solar::factory('Solar_Cli_LinkVendor');
+        $link_vendor->exec($this->_studly);
     }
     
     /**
@@ -241,6 +186,41 @@ class Solar_Cli_MakeVendor extends Solar_Cli_Base
             $text = str_replace('{:php}', '<?php', $text);
             $text = str_replace('{:vendor}', $this->_studly, $text);
             
+            $this->_out("Writing $file ... ");
+            file_put_contents($file, $text);
+            $this->_outln("done.");
+        }
+        
+        // write a "config/default.php" file (empty)
+        $file = "$system/source/{$this->_dashes}/config/default.php";
+        $text = "<?php // placeholder config file for {$this->_studly} classes\n";
+        if (file_exists($file)) {
+            $this->_outln("File $file exists.");
+        } else {
+            $this->_out("Writing $file ... ");
+            file_put_contents($file, $text);
+            $this->_outln("done.");
+        }
+        
+        // write a "config/run-tests.php" file
+        $file = "$system/source/{$this->_dashes}/config/run-tests.php";
+        $text = "<?php\n\$config = array();\nreturn \$config;\n";
+        if (file_exists($file)) {
+            $this->_outln("File $file exists.");
+        } else {
+            $this->_out("Writing $file ... ");
+            file_put_contents($file, $text);
+            $this->_outln("done.");
+        }
+        
+        // write a "vendor/Vendor/Cli/Help/Info/help.txt" file
+        $file = "$system/source/{$this->_dashes}/{$this->_studly}/Cli/Help/Info/help.txt";
+        $text = "{$this->_studly} command line tool." . PHP_EOL . PHP_EOL
+              . "Usage: %*{$this->_dashes} <command> <options> <params>%n" . PHP_EOL . PHP_EOL
+              . "Try '{$this->_dashes} help <command>' for help on a specific command." . PHP_EOL;
+        if (file_exists($file)) {
+            $this->_outln("File $file exists.");
+        } else {
             $this->_out("Writing $file ... ");
             file_put_contents($file, $text);
             $this->_outln("done.");
